requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autograd==1.4
autograd-gamma==0.5.0
autopage==0.5.0
ax-platform==0.2.5.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
botorch==0.6.4
cachetools==5.0.0
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cliff==3.10.1
cloudpickle==2.0.0
cmaes==0.8.2
cmd2==2.4.0
colorlog==6.6.0
cssselect2==0.6.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
DeepPurpose==0.1.5
defusedxml==0.7.1
descriptastorus==2.3.0.6
dgllife==0.2.9
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.6.0
fonttools==4.31.2
formulaic==0.3.4
frozenlist==1.3.0
fsspec==2022.3.0
future==0.18.2
fuzzywuzzy==0.18.0
gdown==4.4.0
gin-config==0.5.0
google-auth==2.6.2
google-auth-oauthlib==0.4.6
gpytorch==1.6.0
greenlet==1.1.2
grpcio==1.44.0
huggingmolecules==0.0.1
hyperopt==0.2.7
idna==3.3
imageio==2.18.0
importlib-metadata==4.11.3
importlib-resources==5.6.0
install==1.3.5
interface-meta==1.3.0
ipykernel==6.12.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.1
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
lazy_loader==0.1rc2
lifelines==0.27.0
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mock==4.0.3
multidict==6.0.2
multipledispatch==0.6.0
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.8
notebook==6.4.10
numpy==1.22.3
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
optuna==2.10.0
packaging==21.3
pandas==1.4.2
pandas-flavor==0.3.0
pandocfilters==1.5.0
parso==0.8.3
pbr==5.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pip==21.2.4
plotly==5.7.0
prettytable==3.2.0
prometheus-client==0.13.1
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
Pygments==2.11.2
pyparsing==3.0.7
pyperclip==1.8.2
pyro-api==0.1.2
pyro-ppl==1.8.0
pyrsistent==0.18.1
PySocks==1.7.1
PyTDC==0.3.6
python-dateutil==2.8.2
pytorch-lightning==1.0.6
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
rdkit-pypi==2022.3.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==0.23.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==58.0.4
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.35
stack-data==0.2.0
stevedore==3.5.0
subword-nmt==0.3.8
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2022.4.28
tinycss2==1.1.1
torch==1.10.1
torch-cluster==1.6.0
torch-geometric==2.0.4
torch-scatter==2.0.9
torch-sparse==0.6.13
torch-spline-conv==1.2.1
torchaudio==0.10.1
torchmetrics==0.7.3
torchvision==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typeguard==2.13.3
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.1
wget==3.2
wheel==0.37.1
widgetsnbextension==3.6.0
wrapt==1.14.1
xarray==2022.3.0
yarl==1.7.2
zipp==3.8.0
